requirements.txt
aiomysql==0.0.20
aioredis==1.3.1
aiosqlite==0.16.0
astroid==2.4.2
async-timeout==3.0.1
autopep8==1.5.4
bcrypt==3.2.0
cffi==1.14.2
click==7.1.2
cryptography==3.1
ecdsa==0.14.1
fastapi==0.62.0
h11==0.11.0
hiredis==1.1.0
httptools==0.1.1
iso8601==0.1.12
isort==5.4.2
lazy-object-proxy==1.4.3
loguru==0.5.1
mccabe==0.6.1
passlib==1.7.2
pep8==1.7.1
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.7.3
pylint==2.6.0
PyMySQL==0.9.2
PyPika==0.38.0
python-jose==3.2.0
python-multipart==0.0.5
redis==3.5.3
rsa==4.6
six==1.15.0
starlette==0.13.6
toml==0.10.1
tortoise-orm==0.16.14
typed-ast==1.4.1
typing-extensions==3.7.4.3
uvicorn==0.12.3
uvloop==0.14.0
websockets==8.1
wrapt==1.12.1
